requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
pytz==2023.3
sqlparse==0.4.4
uvicorn==0.22.0
whitenoise==6.4.0
